Move cognito block
