+++ app/src/js/server/middleware/cognitoAuth.ts
@@ -1,19 +1,26 @@
+import { NextFunction, Request, Response } from 'express'
 import jwt from 'express-jwt'
 import jwksRsa from 'jwks-rsa'
 
 import { ServerConfig } from '../types'
 
-const auth = (config: ServerConfig) =>
-  jwt({
-    secret: jwksRsa.expressJwtSecret({
-      cache: true,
-      rateLimit: true,
-      jwksRequestsPerMinute: 5,
-      jwksUri: `https://cognito-idp.${config.region}.amazonaws.com/${config.userPool}/.well-known/jwks.json`
-    }),
+const auth = (config: ServerConfig) => {
+  if (config.cognito) {
+    return jwt({
+      secret: jwksRsa.expressJwtSecret({
+        cache: true,
+        rateLimit: true,
+        jwksRequestsPerMinute: 5,
+        jwksUri: `https://cognito-idp.${config.cognito.region}.amazonaws.com/${config.cognito.userPool}/.well-known/jwks.json`
+      }),
 
-    issuer: `https://cognito-idp.${config.region}.amazonaws.com/${config.userPool}`,
-    algorithms: [ 'RS256' ]
-  })
+      issuer: `https://cognito-idp.${config.cognito.region}.amazonaws.com/${config.cognito.userPool}`,
+      algorithms: [ 'RS256' ]
+    })
+  } else {
+    return (_request: Request, _response: Response, next: NextFunction) =>
+      next()
+  }
+}
 
 export default auth
