+++ app/src/js/server/util.ts
@@ -56,20 +56,26 @@ export function readConfig (): ServerConfig {
  */
 function validateConfig (config: ServerConfig) {
   if (config.userManagement) {
-    if (!Object.prototype.hasOwnProperty.call(config, 'region')) {
-      throw new Error('Region missed in config ')
-    }
-    if (!Object.prototype.hasOwnProperty.call(config, 'userPool')) {
-      throw new Error('User pool missed in config')
-    }
-    if (!Object.prototype.hasOwnProperty.call(config, 'clientId')) {
-      throw new Error('Client id missed in config')
-    }
-    if (!Object.prototype.hasOwnProperty.call(config, 'userPoolBaseUri')) {
-      throw new Error('User pool base uri missed in config')
-    }
-    if (!Object.prototype.hasOwnProperty.call(config, 'callbackUri')) {
-      throw new Error('Call back uri missed in config')
+    if (config.cognito) {
+      if (!Object.prototype.hasOwnProperty.call(config.cognito, 'region')) {
+        throw new Error('Region missed in config ')
+      }
+      if (!Object.prototype.hasOwnProperty.call(config.cognito, 'userPool')) {
+        throw new Error('User pool missed in config')
+      }
+      if (!Object.prototype.hasOwnProperty.call(config.cognito, 'clientId')) {
+        throw new Error('Client id missed in config')
+      }
+      if (!Object.prototype.hasOwnProperty.call(
+        config.cognito, 'userPoolBaseUri')) {
+        throw new Error('User pool base uri missed in config')
+      }
+      if (!Object.prototype.hasOwnProperty.call(
+        config.cognito, 'callbackUri')) {
+        throw new Error('Call back uri missed in config')
+      }
+    } else {
+      throw new Error('Cognito setting missed in config')
     }
   }
 }
